Fix build issues

+++ src/itmflib/core/itmfunits.h
@@ -95,7 +95,7 @@ namespace itmflib {
 		void setFlags(int32_t f) { flags = BMLint(static_cast<int>(STREAM_HEADER_IDS::FLAGS), f); }
 		void setCodec(int32_t c) { codec = BMLint(static_cast<int>(STREAM_HEADER_IDS::CODEC), c); }
 		void setCodec(std::string c) { codec = BMLstring(static_cast<int>(STREAM_HEADER_IDS::CODEC), c); }
-		void setFormat(std::shared_ptr<char> f, int length) { format = BMLblob(static_cast<int>(STREAM_HEADER_IDS::FORMAT), length, f); }
+		void setFormat(std::shared_ptr<char> f, uint64_t length) { format = BMLblob(static_cast<int>(STREAM_HEADER_IDS::FORMAT), length, f); }
 		void setNumChunks(int64_t n) { nchunks = BMLlong(static_cast<int>(STREAM_HEADER_IDS::NCHUNKS), n); }
 		void addNumChunks(int64_t n) {
 			if (nchunks != boost::none) {
@@ -121,8 +121,8 @@ namespace itmflib {
 		}
 		void setCipher(int32_t c) { cipher = BMLint(static_cast<int>(STREAM_HEADER_IDS::CIPHER), c); }
 		void setCipher(std::string c) { cipher = BMLstring(static_cast<int>(STREAM_HEADER_IDS::CIPHER), c); }
-		void setSalt(std::shared_ptr<char> s, int length) { salt = BMLblob(static_cast<int>(STREAM_HEADER_IDS::SALT), length, s); }
-		void setSignature(std::shared_ptr<char> s, int length) { signature = BMLblob(static_cast<int>(STREAM_HEADER_IDS::SIGNATURE), length, s); }
+		void setSalt(std::shared_ptr<char> s, uint64_t length) { salt = BMLblob(static_cast<int>(STREAM_HEADER_IDS::SALT), length, s); }
+		void setSignature(std::shared_ptr<char> s, uint64_t length) { signature = BMLblob(static_cast<int>(STREAM_HEADER_IDS::SIGNATURE), length, s); }
 
 		// TODO: gets
 		boost::variant<BMLint, BMLstring> getType() const { return type; }
@@ -189,7 +189,7 @@ namespace itmflib {
 		void setNumBytes(BMLlong nb) { nbytes = nb; }
 		void setName(std::string n) { name = BMLstring(static_cast<int>(DIRECTORY_IDS::NAME), n); }
 		void setName(BMLstring n) { name = n; }
-		void setSha256(std::shared_ptr<char> s, int length) { sha256 = BMLblob(static_cast<int>(DIRECTORY_IDS::SHA256), length, s); }
+		void setSha256(std::shared_ptr<char> s, uint64_t length) { sha256 = BMLblob(static_cast<int>(DIRECTORY_IDS::SHA256), length, s); }
 		void setSha256(BMLblob s) { sha256 = s; }
 		void setFlags(int32_t f) { flags = BMLint(static_cast<int>(DIRECTORY_IDS::FLAGS), f); }
 		void setFlags(BMLint f) { flags = f; }
@@ -244,8 +244,8 @@ namespace itmflib {
 		//void setAlgorithm(boost::variant<BMLint, BMLstring> alg) { algorithm = alg; }
 		void setAlgorithm(int32_t alg) { algorithm = BMLint(static_cast<int>(SIGNATURE_IDS::ALGORITHM), alg); }
 		void setAlgorithm(std::string alg) { algorithm = BMLstring(static_cast<int>(SIGNATURE_IDS::ALGORITHM), alg); }
-		void setCertificate(std::shared_ptr<char> c, int length) { certificate = BMLblob(static_cast<int>(SIGNATURE_IDS::CERTIFICATE), length, c); }
-		void setSignature(std::shared_ptr<char> s, int length) { signature = BMLblob(static_cast<int>(SIGNATURE_IDS::SIGNATURE), length, s); }
+		void setCertificate(std::shared_ptr<char> c, uint64_t length) { certificate = BMLblob(static_cast<int>(SIGNATURE_IDS::CERTIFICATE), length, c); }
+		void setSignature(std::shared_ptr<char> s, uint64_t length) { signature = BMLblob(static_cast<int>(SIGNATURE_IDS::SIGNATURE), length, s); }
 
 		size_t write(std::ofstream& outfile);
 		//size_t writeToBuffer(std::vector<char>* buffer);
@@ -273,7 +273,7 @@ namespace itmflib {
 		void setStreamIndex(int32_t si) { stream_index = BMLint(static_cast<int>(CHUNK_IDS::STREAM_INDEX), si); }
 		void setFlags(int32_t f) { flags = BMLint(static_cast<int>(CHUNK_IDS::FLAGS), f); }
 		void setUncompressBytes(int64_t ub) { uncompressed_bytes = BMLlong(static_cast<int>(CHUNK_IDS::UNCOMPRESSED_NBYTES), ub); }
-		void setData(std::shared_ptr<char> d, int length) { data = BMLblob(static_cast<int>(CHUNK_IDS::DATA), length, d); }
+		void setData(std::shared_ptr<char> d, uint64_t length) { data = BMLblob(static_cast<int>(CHUNK_IDS::DATA), length, d); }
 
 		size_t write(std::ofstream& outfile);
 		//size_t writeToBuffer(std::vector<char>* buffer);
